Fix warnings and enable treating warnings as errors (#538)

+++ source/game_sa/Tasks/TaskManager.h
@@ -25,7 +25,7 @@ enum ePrimaryTasks // array indexes
 
 enum eSecondaryTask : uint32 // array indexes
 {
-    TASK_SECONDARY_INVALID = -1,
+    TASK_SECONDARY_INVALID = (uint32)-1,
 
     TASK_SECONDARY_ATTACK = 0,              // want duck to be after attack
     TASK_SECONDARY_DUCK,                    // because attack controls ducking movement
